pyproject.toml
[project]
name = "yolo-talk"
version = "1.0.0"
requires-python = ">=3.12"

dependencies = [
    "fastapi",
    "uvicorn",
    "streamlit",
    "ultralytics",
    "google-generativeai",
    "python-dotenv",
    "pillow",
    "python-multipart",
    "google-genai",
    "dotenv>=0.9.9",
]
